# backend/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Jaya API Documentation",
        "description": "Desafio técnico para a vaga de desenvolvedor na Jaya",
        "contact": {
            "email": "xandemag@gmail.com"
        },
        "license": {
            "name": "MIT License",
            "url": ""
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000",
            "description": "API Server"
        }
    ],
    "paths": {
        "/rest/payments": {
            "get": {
                "tags": [
                    "Payments"
                ],
                "summary": "Get a list of payments",
                "description": "Endpoint to retrieve a list of payments",
                "operationId": "getPayments",
                "responses": {
                    "200": {
                        "description": "List of payments",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "properties": {
                                            "id": {
                                                "type": "integer",
                                                "example": 1
                                            },
                                            "amount": {
                                                "type": "number",
                                                "example": 245.9
                                            },
                                            "status": {
                                                "type": "string",
                                                "example": "completed"
                                            },
                                            "created_at": {
                                                "type": "string",
                                                "format": "date-time",
                                                "example": "2021-08-01T00:00:00.000000Z"
                                            }
                                        },
                                        "type": "object"
                                    }
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "No payments found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "type": "string",
                                            "example": "Not Found"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "No payments found"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Payments"
                ],
                "summary": "Create a new payment",
                "description": "Endpoint to create a new payment",
                "operationId": "createPayment",
                "requestBody": {
                    "description": "Payment details",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "transaction_amount": {
                                        "type": "number",
                                        "example": 245.9
                                    },
                                    "installments": {
                                        "type": "integer",
                                        "example": 3
                                    },
                                    "token": {
                                        "type": "string",
                                        "example": "ae4e50b2a8f3h6d9f2c3a4b5d6e7f8g9"
                                    },
                                    "payment_method_id": {
                                        "type": "string",
                                        "example": "master"
                                    },
                                    "payer": {
                                        "properties": {
                                            "email": {
                                                "type": "string",
                                                "example": "example_random@gmail.com"
                                            },
                                            "identification": {
                                                "properties": {
                                                    "type": {
                                                        "type": "string",
                                                        "example": "CPF"
                                                    },
                                                    "number": {
                                                        "type": "string",
                                                        "example": "12345678909"
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        },
                                        "type": "object"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Payment created successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Payment created successfully"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "type": "string",
                                            "example": "Validation error"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Invalid input data"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/rest/payments/{id}": {
            "get": {
                "tags": [
                    "Payments"
                ],
                "summary": "Get payment details by ID",
                "description": "Endpoint to retrieve details of a payment by ID",
                "operationId": "getPayment",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the payment",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Payment details",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "transaction_amount": {
                                            "type": "number",
                                            "example": 245.9
                                        },
                                        "installments": {
                                            "type": "integer",
                                            "example": 3
                                        },
                                        "token": {
                                            "type": "string",
                                            "example": "ae4e50b2a8f3h6d9f2c3a4b5d6e7f8g9"
                                        },
                                        "payment_method_id": {
                                            "type": "string",
                                            "example": "master"
                                        },
                                        "payer": {
                                            "properties": {
                                                "email": {
                                                    "type": "string",
                                                    "example": "example_random@gmail.com"
                                                },
                                                "identification": {
                                                    "properties": {
                                                        "type": {
                                                            "type": "string",
                                                            "example": "CPF"
                                                        },
                                                        "number": {
                                                            "type": "string",
                                                            "example": "12345678909"
                                                        }
                                                    },
                                                    "type": "object"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Payment not found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "type": "string",
                                            "example": "Not Found"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Payment not found"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            },
            "delete": {
                "tags": [
                    "Payments"
                ],
                "summary": "Cancel payment by ID",
                "description": "Endpoint to cancel a payment by ID",
                "operationId": "cancelPayment",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the payment",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Payment canceled successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Payment canceled successfully"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Payment not found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "type": "string",
                                            "example": "Not Found"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Payment not found"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            },
            "patch": {
                "tags": [
                    "Payments"
                ],
                "summary": "Confirm payment by ID",
                "description": "Endpoint to confirm a payment by updating its status",
                "operationId": "confirmPayment",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the payment",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "description": "Payment confirmation details",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "status": {
                                        "type": "string",
                                        "example": "PAID"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Payment confirmed successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Payment confirmed successfully"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Payment not found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "type": "string",
                                            "example": "Not Found"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Payment not found"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        }
    },
    "tags": [
        {
            "name": "Payments",
            "description": "Operations related to payments"
        }
    ]
}